requirements.txt
requests>=2.28.0
PyYAML>=6.0
pandas>=1.5.0
python-dotenv>=0.20.0
